Revised serialization
