+++ examples/noise_qualification_2d.cpp
@@ -1,7 +1,6 @@
 #include <industrial_calibration/optimizations/analysis/noise_qualification.h>
 #include <industrial_calibration/target_finders/target_finder_plugin.h>
 #include <industrial_calibration/target_finders/utils/utils.h>
-#include <industrial_calibration/serialization/types.h>
 
 #if __GNUC__ >= 8
 #include <filesystem>
