+++ include/industrial_calibration/optimizations/extrinsic_multi_static_camera.h
@@ -14,6 +14,7 @@
 #pragma once
 
 #include <industrial_calibration/types.h>
+#include <industrial_calibration/camera_intrinsics.h>
 #include <industrial_calibration/optimizations/analysis/covariance_types.h>
 
 #include <Eigen/Dense>
