+++ include/industrial_calibration/optimizations/multi_camera_pnp.h
@@ -12,6 +12,7 @@
 #pragma once
 
 #include <industrial_calibration/types.h>
+#include <industrial_calibration/camera_intrinsics.h>
 
 namespace industrial_calibration
 {
