+++ include/industrial_calibration/optimizations/pnp.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include <industrial_calibration/optimizations/analysis/covariance_types.h>
 #include <industrial_calibration/types.h>
+#include <industrial_calibration/camera_intrinsics.h>
+#include <industrial_calibration/optimizations/analysis/covariance_types.h>
 
 namespace industrial_calibration
 {
